requirements.txt
# Core ML & NLP
transformers==4.39.2
sentence-transformers==2.7.0 
torch==2.2.1
llama-cpp-python==0.2.57

# Data Handling & Plotting
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2